+++ packages/expo-modules-core/ios/Worklets/WorkletJSCallInvoker.cpp
@@ -25,10 +25,7 @@ void WorkletJSCallInvoker::invokeSync(react::CallFunc &&func) {
     return;
   }
 
-  workletRuntime->executeSync([func = std::move(func)](jsi::Runtime &rt) -> jsi::Value {
-    func(rt);
-    return jsi::Value::undefined();
-  });
+  workletRuntime->runSync(func);
 }
 
 } // namespace expo